Number Enabler Interface
I did a few interfaces for the number so that the user can enable the number when they want. I'm not sure if you guys want to me to run a loop around the cells so it is read only? Then the user can manually select a number and after they select a cell, the number inside will be replaced.