Validation and "Winnable"
Check Answers button was used to validate answers given. If all is correct you have won the game! I check that with a bool. If you have wrong answers those wrong answers will give the texbox a red background. Added a line on the GameCellTextChange method to turn the background back to the default and gets rid of the red.